Break all lambda logic out into separate files
