+++ handlers/supporter-product-data-lambdas/src/lambdas/addSupporterRatePlanItemToQueueLambda.ts
@@ -1,24 +1,17 @@
 import { sendBatchMessagesToQueue } from '@modules/aws/sqs';
 import { Lazy } from '@modules/lazy';
 import { stageFromEnvironment } from '@modules/stage';
-import type { SupporterRatePlanItem } from '@modules/supporter-product-data/supporterProductData';
 import type { Handler } from 'aws-lambda';
-import { addToQueue } from '../services/addItemsToQueue';
+import {
+	addToQueue,
+	type AddToQueueDependencies,
+} from '../services/addItemsToQueue';
 import { ConfigService } from '../services/configService';
 import { parseCsvStreamWithHeader } from '../services/csvService';
 import { S3Service } from '../services/s3Service';
 import type { AddSupporterRatePlanItemToQueueState } from './types';
 
 const stage = stageFromEnvironment();
-type IndexedItem = [SupporterRatePlanItem, number];
-
-export type AddToQueueDependencies = {
-	streamCsvRows: (
-		filename: string,
-	) => AsyncIterable<Record<string, string>> | Iterable<Record<string, string>>;
-	sendMessagesToQueue: (items: IndexedItem[]) => Promise<void>;
-	putLastSuccessfulQueryTime: (time: string) => Promise<void>;
-};
 
 const buildDependencies = (): AddToQueueDependencies => {
 	const s3Service = new S3Service();
