+++ handlers/supporter-product-data-lambdas/src/services/contributions.ts
@@ -1,7 +1,7 @@
 import { logger } from '@modules/logger/logger';
 import type { Stage } from '@modules/stage';
 import type { SupporterRatePlanItem } from '@modules/supporter-product-data/supporterProductData';
-import type { ProcessItemDependencies } from '../lambdas/processSupporterRatePlanItemLambda';
+import type { ProcessItemDependencies } from './processSupporterRatePlanItem';
 import type { MinimalZuoraSubscription } from './zuoraSubscriptionService';
 
 export const contributionIdsForStage = (stage: Stage): string[] =>
